Verus now supports recursive-submodule verif
Verus#385 is basically resolved now, so we don't need the harsh warning
and workaround and only need it if someone is actually hitting a bad
case within single-function verification.

Basically, this should make it such that files that have submodules can
get only that module and submodules.  It also fixes up the issue with
single-function-verification that showed up once Verus changed the
meaning of `--verify-module`.