Parse secondary spans from Verus
The shortlog was better in the past because the spans otherwise would
lead to some pretty undesirable behavior in terms of red-squigglies
everywhere. However, Verus (earlier this year) updated its span
information, meaning that the JSON log is actually quite reasonable.
Thus, by parsing that, we can get secondary span information too, which
is helpful to provide more accurate squigglies in flycheck mode.